+++ CHANGELOG.md
@@ -40,6 +40,8 @@
         - allow jsonb values properly
         - create empty tables with strange dump characters
         - truncation on json rows wider than 1mb
+      - (redis) hard error on ALL
+      - (redis) allow name: int for backup and restore
       - (storage) fixes
         - (azure/s3/webdav) open new handle on each connection attempt
         - (s3) default to https if url is not passed